refactor game entity related in reload component to a utility class
